+++ tests/connectors.py
@@ -4,50 +4,89 @@
 
 import os
 from typing import Dict, List
-import pathlib
 from meerschaum import get_connector
 from meerschaum.config._paths import ROOT_DIR_PATH
 data_path = ROOT_DIR_PATH / 'data'
 data_path.mkdir(exist_ok=True)
 
 conns = {
-    'timescaledb': get_connector('sql', 'test_timescaledb',
-        flavor='timescaledb', username='test', password='test1234', database='testdb',
-        port=5439, host='localhost', schema='public',
+    'timescaledb': get_connector(
+        'sql', 'test_timescaledb',
+        flavor='timescaledb',
+        username='test',
+        password='test1234',
+        database='testdb',
+        port=5439,
+        host='localhost',
+        schema='public',
     ),
-    'mariadb': get_connector('sql', 'test_mariadb',
-        flavor='mariadb', username='test', password='test1234', database='testdb',
-        port=3309, host='localhost',
+    'mariadb': get_connector(
+        'sql', 'test_mariadb',
+        flavor='mariadb',
+        username='test',
+        password='test1234',
+        database='testdb',
+        port=3309,
+        host='localhost',
     ),
-    'mysql': get_connector('sql', 'test_mysql',
-        flavor='mysql', username='root', password='my-secret-pw', database='mysql',
-        port=3310, host='localhost',
+    'mysql': get_connector(
+        'sql', 'test_mysql',
+        flavor='mysql',
+        username='root',
+        password='my-secret-pw',
+        database='mysql',
+        port=3310,
+        host='localhost',
     ),
-    'mssql': get_connector('sql', 'test_mssql',
-        flavor='mssql', username='sa', password='supersecureSECRETPASSWORD123!',
-        database='master', port=1439, host='localhost', schema='dbo',
+    'mssql': get_connector(
+        'sql', 'test_mssql',
+        flavor='mssql',
+        username='sa',
+        password='supersecureSECRETPASSWORD123!',
+        database='master',
+        port=1439,
+        host='localhost',
+        schema='dbo',
     ),
-    'oracle': get_connector('sql', 'test_oracle',
-        flavor='oracle', host='localhost', database='xe', username='system', password='oracle',
+    'oracle': get_connector(
+        'sql', 'test_oracle',
+        flavor='oracle',
+        host='localhost',
+        database='xe',
+        username='system',
+        password='oracle',
         port=1529,
     ),
-    'sqlite':  get_connector('sql', 'test_sqlite',
+    'sqlite':  get_connector(
+        'sql', 'test_sqlite',
         database=str(data_path / 'test_sqlite.db'),
         flavor='sqlite',
     ),
-    'duckdb': get_connector('sql', 'test_duckdb',
+    'duckdb': get_connector(
+        'sql', 'test_duckdb',
         database=str(data_path / 'test_duck.db'),
         flavor='duckdb',
     ),
-    'citus': get_connector('sql', 'test_citus',
-        flavor='citus', username='test', password='test1234', database='testdb',
-        port=5499, host='localhost',
+    'citus': get_connector(
+        'sql', 'test_citus',
+        flavor='citus',
+        username='test',
+        password='test1234',
+        database='testdb',
+        port=5499,
+        host='localhost',
     ),
-    'api': get_connector('api', 'test_api',
-        port=8989, username='test', password='test1234', host='localhost',
+    'api': get_connector(
+        'api', 'test_api',
+        port=8989,
+        username='test',
+        password='test1234',
+        host='localhost',
     ),
-    'valkey': get_connector('valkey', 'test_valkey',
-        port=6399, host='localhost',
+    'valkey': get_connector(
+        'valkey', 'test_valkey',
+        port=6399,
+        host='localhost',
     ),
 }
 
@@ -66,7 +105,7 @@ def get_dtypes(debug: bool = False) -> Dict[str, Dict[str, 'sqlalchemy.types.Typ
             'value': '2022-01-01',
         },
         'datetime64[ns, UTC]': {
-            'value': '2022-01-01',
+            'value': '2022-01-01 00:00:00+00:00',
         },
         'float': {
             'value': 1.0,